rename TopActivity to MainActivity
