+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/RepoViewFragment.kt
@@ -9,7 +9,7 @@ import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.navArgs
 import coil.load
-import jp.co.yumemi.android.codecheck.TopActivity.Companion.lastSearchDate
+import jp.co.yumemi.android.codecheck.MainActivity.Companion.lastSearchDate
 import jp.co.yumemi.android.codecheck.databinding.RepoViewFragmentBinding
 
 class RepoViewFragment : Fragment(R.layout.repo_view_fragment) {
